Move the CardField/CardFieldRow/CardFieldDivider trio out of StakePositionsTable into components/product/CardFields (PR review): the mobile-card field primitives sit next to TransactionCard with the other shared product pieces instead of StakeActivityTable importing them from a sibling table
